Added RRT. Setting up constellation generation!
RRT stands for Rapidly-Exploring Random Trees and is an
algorthm used in robotics. It creates a nice looking tree
that expands to fill a space.

This commit is mostly to get the base version in place.
The node connections and data-structure still needs to be
added in. This requires some modifications. So this commit
is a way to preserve this pure form of this alogorithm.