Added a ConstellationSpawner Prefab
Updated the constellation spawner to support genre colors.
DataContainer currently contains a test script that I used to debug the

constellation spawning. I plan on replacing it with a proper manager
later.

Updated NormalNode to have more consistant behavior and to allow for
sizes to be set.

ConstellationSpawner Prefabs will be used to generate the actual
constellations. They should be fine for now, may need to extend / tweak
them a bit.

The CSV reader is completely broken for this dataset, will update soon.
Next up is adding controls that allow the player to view node
information.