Fixed the drawing/selection issues. Added feedback
Updated the assests to improve beam color.

Implemented a temp color change for nodes to allow them to be
highlighted.

Fixed a bug that caused the beam to flicker when used. This was caused
by by the raycaster colliding with the controller models. I fixed this
by creating a seperate rendering layer just for the nodes. This allows
the raycaster to go through every other object. This also stops the
raycaster from not selecting an object it's pointed at...

Updated the control manisfest to fix a bug that caused all the bindings
to reset?

Updated normal node to have a color set function.

Removed the emission from the LineMaterial, this allows for the fade in.